+++ src/cljs/rems/auth/auth.cljs
@@ -1,6 +1,5 @@
 (ns rems.auth.auth
   (:require [re-frame.core :as rf]
-            [rems.auth.ldap :as ldap]
             [rems.auth.oidc :as oidc]
             [rems.auth.shibboleth :as shibboleth])
   (:require-macros [rems.guide-macros :refer [component-info example]]))
@@ -11,7 +10,6 @@
         login-component (case (:authentication config)
                           :shibboleth (shibboleth/login-component alternative-endpoint)
                           :fake-shibboleth (shibboleth/login-component alternative-endpoint)
-                          :ldap (ldap/login-component)
                           :oidc (oidc/login-component)
                           nil)]
     (when login-component
@@ -23,7 +21,5 @@
    (component-info shibboleth/login-component)
    (example "shibboleth login" [shibboleth/login-component nil])
    (example "shibboleth login with alternatives" [shibboleth/login-component "/alternative"])
-   (component-info ldap/login-component)
-   (example "ldap login" [ldap/login-component])
    (component-info oidc/login-component)
    (example "oidc login" [oidc/login-component])])