+++ CHANGELOG.md
@@ -9,10 +9,11 @@ have notable changes.
 Changes since v2.6
 
 ### Breaking changes
+- Removed support for LDAP authentication
 - `/api/workflows/create` API: the `type` parameter's allowed value was changed from `dynamic` to `workflow/dynamic`
 
 ### Additions
-- Catalogue item form can be changed for one or more items at a time. 
+- Catalogue item form can be changed for one or more items at a time.
   New items will be created that use the new form while the old items
   are disabled and archived. The name of the new item will be exactly
   the same as before. See #837
@@ -22,6 +23,7 @@ Changes since v2.6
 - Redirect to login page when accessing an attachment link when logged out (#1590)
 - Form editor: add new field between fields (#1812)
 - Entitlements appear immediately instead of after a delay (#1784)
+- Show version information in console instead of the page footer (#1785)
 
 ### Fixes
 - More robust email resending (#1750)